FIX: move event_sale code to event_sale
